[#12] Working grammar code, rust fmt, refactoring
This is a squashed commit because I accidentally commited a big file.
It was easier to squash a few recent commits than to remove it from the
history.

The grammar code was made functional after lots of changes and
debugging. This code is slow and all round not very good. This is a
start to the refactoring